MicroServices/FeaturePipeline/KafkaToFeatureStore/pyproject.toml
[tool.poetry]
name = "source"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.11, <3.13"
quixstreams = "^2.8.1"
loguru = "^0.7.2"
python-dotenv = "^1.0.1"
hopsworks = "^3.7.0"

[tool.poetry.group.dev.dependencies]
ruff = "^0.5.7"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.ruff]
line-length = 88

[tool.ruff.format]
quote-style = "single"
indent-style = "space"
docstring-code-format = true

[tool.ruff.lint]
extend-select = ["I"]